- Added GTMNSMakeUncollectable for forcing objects to survive in a GC world.
- Added GTMCFAutorelease to make the [GTMNSMakeCollectable(cfFoo) autorelease] simpler and clearer, it's now just GTMCFAutorelease(cfFoo), and works in both GC and non-GC world.

- Added GTMIsGarbageCollectionEnabled to GTMGarbageCollection.h.  See the note there for it's usage.

- Disabled the unittests for things on top of NSAppleScript in a GC world since Apple has bugs and it can crash.  See the unittest for a note about it.

- GTMStackTrace now can figure out ObjC symbols. Downside it is now ObjC only.

- GTMFourCharCode can now be used with NSAppleEventDescriptors easily. typeType, typeKeyword, typeApplSignature, and typeEnumerated all get turned into GTMFourCharCodes.

- Fixed up crash in GTMLoggerRingBufferWriter when used with GC on.

- Significant updates to GTMNSAppleScript+Handler allowing you to list all handlers and properties (including inherited) and cleans up several errors in how scripting was being handled.

+++ trunk/Foundation/GTMLoggerRingBufferWriter.h
@@ -17,6 +17,7 @@
 //
 
 #import "GTMLogger.h"
+#import "GTMDefines.h"
 
 typedef struct GTMRingBufferPair GTMRingBufferPair;
 
@@ -51,39 +52,39 @@ typedef struct GTMRingBufferPair GTMRingBufferPair;
  @private  
   id<GTMLogWriter> writer_;
   GTMRingBufferPair *buffer_;
-  int capacity_;
-  int nextIndex_;    // Index of the next element of |buffer_| to fill.
-  int totalLogged_;  // This > 0 and |nextIndex_| == 0 means we've wrapped.
+  NSUInteger capacity_;
+  NSUInteger nextIndex_;    // Index of the next element of |buffer_| to fill.
+  NSUInteger totalLogged_;  // This > 0 and |nextIndex_| == 0 means we've wrapped.
 }
 
-// Returns an autoreleased ring buffer writer.  If |capacity| is
-// non-positive, or |writer| is nil, then nil is returned.
-+ (id)ringBufferWriterWithCapacity:(int)capacity
+// Returns an autoreleased ring buffer writer.  If |writer| is nil, 
+// then nil is returned.
++ (id)ringBufferWriterWithCapacity:(NSUInteger)capacity
                             writer:(id<GTMLogWriter>)loggerWriter;
 
-// Designated initializer.  If |capacity| is non-positive, or |writer|
-// is nil, then nil is returned.  If you just use -init, nil will be returned.
-- (id)initWithCapacity:(int)capacity
+// Designated initializer.  If |writer| is nil, then nil is returned.  
+// If you just use -init, nil will be returned.
+- (id)initWithCapacity:(NSUInteger)capacity
                 writer:(id<GTMLogWriter>)loggerWriter;
 
 // How many messages will be logged before older messages get dropped
 // on the floor.
-- (int)capacity;
+- (NSUInteger)capacity;
 
 // The log writer that will get the buffered log messages if/when they
 // need to be displayed.
 - (id<GTMLogWriter>)writer;
 
 // How many log messages are currently in the buffer.
-- (int)count;
+- (NSUInteger)count;
 
 // How many have been dropped on the floor since creation, or the last
 // reset.
-- (int)droppedLogCount;
+- (NSUInteger)droppedLogCount;
 
 // The total number of messages processed since creation, or the last
 // reset.
-- (int)totalLogged;
+- (NSUInteger)totalLogged;
 
 // Purge the contents and reset the counters.
 - (void)reset;
